Remove ES6 features. Wrap the code in a strict-mode function in preparation for browser usage. Add comments.
